Add actual tests, fix a few bugs
Consider a 'n' in the type specifer as integer if the unsigned flag is
given.
The uppercase check for the 'f' and 'e' specifier now actually checks
the type character, instead of the closing '}'.
The '%' format specifier now multiplies the number by 100.
The 'n' format specifier no longer tries to get it's argument a second
time.
The 'e' format specifier now uses the alternate form for the base.