new float printing implementation
Completely new float printing implementation using the Grisu2 algorithm,
this is about 4x faster than the old implementation and does not require
linking the math library.  This also comes with a new mode for printing
floats where the exponential mode is used if the exponent is less than
-6 or greater than 20 and decimal is used by default, this is also the
default mode for floats now.  Furthermore specifying -1 as the default
precision for floats now means that their fraction is not limited at all
and if it's zero is omitted, this has also become the new default.