BREAKING CHANGE: redefine string precision
String precision now comes in 3 flavors, units, null-terminated
codepoints, and absolute codepoints; previously it was always
absolute codepoints.
I'm aware that this is extremely undertested and this version is not
unlikely to contain bugs but I've added this because I need it in
another project and I can't be bothered to extend testing right now.